+++ src/config.rs
@@ -8,6 +8,9 @@ use serde::Deserialize;
 
 use crate::identity::{IdtAmount, UserAddress};
 
+pub const DEFAULT_CONFIG_PATH: &str = "config.json";
+pub const DEFAULT_GENESIS_PATH: &str = "genesis.json";
+
 #[derive(Debug, Clone, Deserialize, Default)]
 pub struct AdminsSection {
     #[serde(default)]
@@ -16,28 +19,12 @@ pub struct AdminsSection {
     pub moderators: HashSet<String>,
 }
 
-#[derive(Debug, Clone, Deserialize, PartialEq, Eq)]
-pub struct ExternalServer {
-    pub url: String,
-    #[serde(default)]
-    pub alias: Option<String>,
-}
-
-#[derive(Debug, Clone, Deserialize)]
+#[derive(Debug, Clone, Deserialize, Default)]
 pub struct ExternalServersSection {
     #[serde(default)]
     pub allow_all: bool,
     #[serde(default)]
-    pub servers: Vec<ExternalServer>,
-}
-
-impl Default for ExternalServersSection {
-    fn default() -> Self {
-        Self {
-            allow_all: false,
-            servers: Vec::new(),
-        }
-    }
+    pub allow_servers: HashSet<UserAddress>,
 }
 
 #[derive(Debug, Clone, Deserialize, Default)]
@@ -48,14 +35,11 @@ pub struct Config {
     pub external_servers: ExternalServersSection,
 }
 
-pub const DEFAULT_CONFIG_PATH: &str = "config.json";
-pub const DEFAULT_GENESIS_PATH: &str = "genesis.json";
-
 pub async fn load_config(path: &str) -> Result<Config, io::Error> {
     let content = match fs::read_to_string(path).await {
         Ok(content) => content,
         Err(err) => {
-            log::warn!("Failed to read {path}: {}", err);
+            log::warn!("Failed to read {}: {}", path, err);
             return Ok(Config::default());
         }
     };
@@ -67,7 +51,7 @@ pub async fn load_genesis(path: &str) -> Result<HashMap<UserAddress, IdtAmount>,
     let content = match fs::read_to_string(path).await {
         Ok(content) => content,
         Err(err) => {
-            log::warn!("Failed to read {path}: {}", err);
+            log::warn!("Failed to read {}: {}", path, err);
             return Ok(HashMap::new());
         }
     };
@@ -77,7 +61,26 @@ pub async fn load_genesis(path: &str) -> Result<HashMap<UserAddress, IdtAmount>,
 
 #[cfg(test)]
 mod tests {
+    use std::path::PathBuf;
+
+    use async_std::{fs::File, io::WriteExt};
+    use tempdir::TempDir;
+
     use super::*;
+
+    async fn create_test_config(dir: &TempDir, content: &str) -> PathBuf {
+        let path = dir.path().join("config.json");
+        let mut file = File::create(&path).await.unwrap();
+        file.write_all(content.as_bytes()).await.unwrap();
+        path
+    }
+
+    async fn create_test_genesis_config(dir: &TempDir, content: &str) -> PathBuf {
+        let path = dir.path().join("genesis.json");
+        let mut file = File::create(&path).await.unwrap();
+        file.write_all(content.as_bytes()).await.unwrap();
+        path
+    }
 
     #[test]
     fn test_parse_default() {
@@ -86,6 +89,7 @@ mod tests {
         assert!(cfg.admins.admins.is_empty());
         assert!(cfg.admins.moderators.is_empty());
         assert!(!cfg.external_servers.allow_all);
+        assert!(cfg.external_servers.allow_servers.is_empty());
     }
 
     #[test]
@@ -94,16 +98,90 @@ mod tests {
             "admins": {"admins": ["a"], "moderators": []},
             "external_servers": {
                 "allow_all": false,
-                "servers": [
-                    {"url": "http://a.com", "alias": "a"},
-                    {"url": "http://b.com"}
-                ]
+                "allow_servers": ["address1", "address2"]
             }
         }"#;
         let cfg: Config = serde_json::from_str(json).unwrap();
         assert_eq!(cfg.admins.admins.len(), 1);
-        assert_eq!(cfg.external_servers.servers.len(), 2);
-        assert_eq!(cfg.external_servers.servers[0].alias.as_deref(), Some("a"));
-        assert_eq!(cfg.external_servers.servers[1].alias, None);
+        assert_eq!(cfg.external_servers.allow_servers.len(), 2);
+        assert!(
+            cfg.external_servers
+                .allow_servers
+                .contains(&"address1".to_string())
+        );
+        assert!(
+            cfg.external_servers
+                .allow_servers
+                .contains(&"address2".to_string())
+        );
+    }
+
+    #[async_std::test]
+    async fn test_load_config_nonexistent_file() {
+        let temp_dir = TempDir::new("config").unwrap();
+        let path = temp_dir.path().join("nonexistent.json");
+        let cfg = load_config(path.to_str().unwrap()).await.unwrap();
+        assert!(cfg.admins.admins.is_empty());
+        assert!(cfg.admins.moderators.is_empty());
+        assert!(cfg.external_servers.allow_servers.is_empty());
+    }
+
+    #[async_std::test]
+    async fn test_load_config_valid_file() {
+        let temp_dir = TempDir::new("config").unwrap();
+        let content = r#"{
+            "admins": {"admins": ["user"], "moderators": ["mod"]},
+            "external_servers": {"allow_all": true, "allow_servers": ["address1"]}
+        }"#;
+        let path = create_test_config(&temp_dir, content).await;
+        let cfg = load_config(path.to_str().unwrap()).await.unwrap();
+        assert_eq!(cfg.admins.admins.len(), 1);
+        assert_eq!(cfg.admins.moderators.len(), 1);
+        assert!(cfg.external_servers.allow_all);
+        assert_eq!(cfg.external_servers.allow_servers.len(), 1);
+    }
+
+    #[async_std::test]
+    async fn test_load_config_invalid_json() {
+        let temp_dir = TempDir::new("config").unwrap();
+        let path = create_test_config(&temp_dir, "{").await;
+        assert!(load_config(path.to_str().unwrap()).await.is_err());
+    }
+
+    #[async_std::test]
+    async fn test_load_config_empty_file() {
+        let temp_dir = TempDir::new("config").unwrap();
+        let path = create_test_config(&temp_dir, "").await;
+        assert!(load_config(path.to_str().unwrap()).await.is_err());
+    }
+
+    #[async_std::test]
+    async fn test_load_genesis_nonexistent_file() {
+        let temp_dir = TempDir::new("config").unwrap();
+        let path = temp_dir.path().join("nonexistent.json");
+        let balances = load_genesis(path.to_str().unwrap()).await.unwrap();
+        assert!(balances.is_empty());
+    }
+
+    #[async_std::test]
+    async fn test_load_genesis_valid() {
+        let temp_dir = TempDir::new("config").unwrap();
+        let path = create_test_genesis_config(&temp_dir, "{\"alice\":1}").await;
+        let balances = load_genesis(path.to_str().unwrap()).await.unwrap();
+        assert_eq!(balances["alice"], 1);
+    }
+
+    #[async_std::test]
+    async fn test_load_genesis_invalid() {
+        let temp_dir = TempDir::new("config").unwrap();
+        let path = create_test_genesis_config(&temp_dir, "{").await;
+        assert!(load_genesis(path.to_str().unwrap()).await.is_err());
+    }
+
+    #[async_std::test]
+    async fn test_load_genesis_empty() {
+        let temp_dir = TempDir::new("config").unwrap();
+        let path = create_test_genesis_config(&temp_dir, "").await;
+        assert!(load_genesis(path.to_str().unwrap()).await.is_err());
     }
 }
