+++ src/lib.rs
@@ -2,4 +2,5 @@ pub mod admins;
 pub mod config;
 pub mod identity;
 pub mod routes;
+pub mod storage;
 pub mod verify;
