+++ src/main.rs
@@ -1,38 +1,18 @@
 use std::{
     env,
-    io::{Error, ErrorKind, Write},
-    sync::Arc,
+    io::{Error, Write},
 };
 
 use identity_server::{
-    admins::{AdminStorage, db::DatabaseAdminStorage},
     config::{self, DEFAULT_CONFIG_PATH, DEFAULT_GENESIS_PATH},
-    identity::{
-        IdentityService, IdtAmount, UserAddress,
-        proof::{db::DatabaseProofStorage, storage::ProofStorage},
-        punish::{db::DatabasePenaltyStorage, storage::PenaltyStorage},
-        vouch::{db::DatabaseVouchStorage, storage::VouchStorage},
-    },
+    identity::IdentityService,
     routes::{self, State},
-    verify::nonce::{NonceManager, db::DatabaseNonceManager},
+    storage,
 };
 use tide::Server;
 
 pub const DEFAULT_PORT: u32 = 8080;
 pub const DEFAULT_HOST: &str = "localhost";
-
-pub const DEFAULT_MYSQL_USER: &str = "root";
-pub const DEFAULT_MYSQL_HOST: &str = "localhost";
-pub const DEFAULT_MYSQL_PORT: u32 = 3306;
-pub const DEFAULT_MYSQL_DATABASE: &str = "identity";
-
-struct Storage {
-    vouch_storage: Arc<dyn VouchStorage>,
-    proof_storage: Arc<dyn ProofStorage>,
-    penalty_storage: Arc<dyn PenaltyStorage>,
-    admin_storage: Arc<dyn AdminStorage>,
-    nonce_manager: Arc<dyn NonceManager>,
-}
 
 #[async_std::main]
 async fn main() {
@@ -47,7 +27,7 @@ async fn main() {
             writeln!(buf, "[{}] {}: {}", record.level(), ts, record.args())
         })
         .init();
-    let mut config = match config::load_config(DEFAULT_CONFIG_PATH).await {
+    let config = match config::load_config(DEFAULT_CONFIG_PATH).await {
         Ok(config) => config,
         Err(e) => {
             log::error!("Failed to load configuration: {:?}", e);
@@ -61,14 +41,14 @@ async fn main() {
             panic!("Failed to load genesis configuration: {}", e);
         }
     };
-    let _external_servers = config.external_servers.clone();
-    let storage = match create_storage(config.admins).await {
-        Ok(storage) => storage,
-        Err(e) => {
-            log::error!("Failed to connect to database: {:?}", e);
-            panic!("Failed to connect to database: {}", e);
-        }
-    };
+    let storage =
+        match storage::create_storage(config.admins.admins, config.admins.moderators).await {
+            Ok(storage) => storage,
+            Err(e) => {
+                log::error!("Failed to connect to database: {:?}", e);
+                panic!("Failed to connect to database: {}", e);
+            }
+        };
 
     let identity_service = IdentityService {
         vouches: storage.vouch_storage,
@@ -94,54 +74,6 @@ async fn main() {
         log::error!("Failed to start server: {:?}", err);
         panic!("Failed to start server: {}", err);
     }
-}
-
-async fn create_storage(config: config::AdminsSection) -> Result<Storage, Error> {
-    let db_url = setup_database_url();
-    let vouch_storage_connect = DatabaseVouchStorage::new(&db_url)
-        .await
-        .map_err(|e| Error::new(ErrorKind::NotConnected, e.to_string()))?;
-    let proof_storage_connect = DatabaseProofStorage::new(&db_url)
-        .await
-        .map_err(|e| Error::new(ErrorKind::NotConnected, e.to_string()))?;
-    let penalty_storage_connect = DatabasePenaltyStorage::new(&db_url)
-        .await
-        .map_err(|e| Error::new(ErrorKind::NotConnected, e.to_string()))?;
-    let admin_storage_connect =
-        DatabaseAdminStorage::new(&db_url, config.admins, config.moderators)
-            .await
-            .map_err(|e| Error::new(ErrorKind::NotConnected, e.to_string()))?;
-    let nonce_manager = DatabaseNonceManager::new(&db_url)
-        .await
-        .map_err(|e| Error::new(ErrorKind::NotConnected, e.to_string()))?;
-    Ok(Storage {
-        vouch_storage: Arc::new(vouch_storage_connect),
-        proof_storage: Arc::new(proof_storage_connect),
-        penalty_storage: Arc::new(penalty_storage_connect),
-        admin_storage: Arc::new(admin_storage_connect),
-        nonce_manager: Arc::new(nonce_manager),
-    })
-}
-
-fn setup_database_url() -> String {
-    let db_user = match env::var("MYSQL_USER").unwrap_or_default().as_str() {
-        "" => DEFAULT_MYSQL_USER.to_string(),
-        user_str => user_str.to_string(),
-    };
-    let db_password = env::var("MYSQL_PASSWORD").unwrap_or_default();
-    let db_host = match env::var("MYSQL_HOST").unwrap_or_default().as_str() {
-        "" => DEFAULT_MYSQL_HOST.to_string(),
-        host_str => host_str.to_string(),
-    };
-    let db_port = match env::var("MYSQL_PORT").unwrap_or_default().as_str() {
-        "" => DEFAULT_MYSQL_PORT,
-        port_str => port_str.parse::<u32>().unwrap_or(DEFAULT_MYSQL_PORT),
-    };
-    let db_name = match env::var("MYSQL_DATABASE").unwrap_or_default().as_str() {
-        "" => DEFAULT_MYSQL_DATABASE.to_string(),
-        db_str => db_str.to_string(),
-    };
-    format!("mysql://{db_user}:{db_password}@{db_host}:{db_port}/{db_name}")
 }
 
 async fn start_server(state: State) -> Result<(), Error> {
@@ -181,93 +113,4 @@ async fn setup_routes(server: &mut Server<State>) {
     server
         .at("/remove_moderator/:user")
         .post(routes::admins::remove_moderator::route);
-}
-#[cfg(test)]
-mod tests {
-    use super::*;
-    use async_std::{fs::File, io::WriteExt};
-    use std::path::PathBuf;
-    use tempdir::TempDir;
-
-    async fn create_test_config(dir: &TempDir, content: &str) -> PathBuf {
-        let path = dir.path().join("config.json");
-        let mut file = File::create(&path).await.unwrap();
-        file.write_all(content.as_bytes()).await.unwrap();
-        path
-    }
-
-    async fn create_test_genesis_config(dir: &TempDir, content: &str) -> PathBuf {
-        let path = dir.path().join("genesis.json");
-        let mut file = File::create(&path).await.unwrap();
-        file.write_all(content.as_bytes()).await.unwrap();
-        path
-    }
-
-    #[async_std::test]
-    async fn test_load_config_nonexistent_file() {
-        let temp_dir = TempDir::new("config").unwrap();
-        let path = temp_dir.path().join("nonexistent.json");
-        let cfg = config::load_config(path.to_str().unwrap()).await.unwrap();
-        assert!(cfg.admins.admins.is_empty());
-        assert!(cfg.admins.moderators.is_empty());
-        assert!(cfg.external_servers.servers.is_empty());
-    }
-
-    #[async_std::test]
-    async fn test_load_config_valid_file() {
-        let temp_dir = TempDir::new("config").unwrap();
-        let content = r#"{
-            "admins": {"admins": ["user"], "moderators": ["mod"]},
-            "external_servers": {"allow_all": false, "servers": [{"url": "http://a.com"}]}
-        }"#;
-        let path = create_test_config(&temp_dir, content).await;
-        let cfg = config::load_config(path.to_str().unwrap()).await.unwrap();
-        assert_eq!(cfg.admins.admins.len(), 1);
-        assert_eq!(cfg.admins.moderators.len(), 1);
-        assert_eq!(cfg.external_servers.servers.len(), 1);
-    }
-
-    #[async_std::test]
-    async fn test_load_config_invalid_json() {
-        let temp_dir = TempDir::new("config").unwrap();
-        let path = create_test_config(&temp_dir, "{").await;
-        assert!(config::load_config(path.to_str().unwrap()).await.is_err());
-    }
-
-    #[async_std::test]
-    async fn test_load_config_empty_file() {
-        let temp_dir = TempDir::new("config").unwrap();
-        let path = create_test_config(&temp_dir, "").await;
-        assert!(config::load_config(path.to_str().unwrap()).await.is_err());
-    }
-
-    #[async_std::test]
-    async fn test_load_genesis_nonexistent_file() {
-        let temp_dir = TempDir::new("config").unwrap();
-        let path = temp_dir.path().join("nonexistent.json");
-        let balances = config::load_genesis(path.to_str().unwrap()).await.unwrap();
-        assert!(balances.is_empty());
-    }
-
-    #[async_std::test]
-    async fn test_load_genesis_valid() {
-        let temp_dir = TempDir::new("config").unwrap();
-        let path = create_test_genesis_config(&temp_dir, "{\"alice\":1}").await;
-        let balances = config::load_genesis(path.to_str().unwrap()).await.unwrap();
-        assert_eq!(balances["alice"], 1);
-    }
-
-    #[async_std::test]
-    async fn test_load_genesis_invalid() {
-        let temp_dir = TempDir::new("config").unwrap();
-        let path = create_test_genesis_config(&temp_dir, "{").await;
-        assert!(config::load_genesis(path.to_str().unwrap()).await.is_err());
-    }
-
-    #[async_std::test]
-    async fn test_load_genesis_empty() {
-        let temp_dir = TempDir::new("config").unwrap();
-        let path = create_test_genesis_config(&temp_dir, "").await;
-        assert!(config::load_genesis(path.to_str().unwrap()).await.is_err());
-    }
 }
